set up shared path
