+++ server/api/puzzle.ts
@@ -1,17 +1,9 @@
+import {AddPuzzleResponse, AddPuzzleRequest} from '@shared/types';
 import express from 'express';
 
-import {addPuzzle, PuzzleJson} from '../model/puzzle';
+import {addPuzzle} from '../model/puzzle';
 
 const router = express.Router();
-
-interface AddPuzzleRequest {
-  puzzle: PuzzleJson;
-  isPublic: boolean;
-}
-
-interface AddPuzzleResponse {
-  pid: string;
-}
 
 router.post<{}, AddPuzzleResponse, AddPuzzleRequest>('/', async (req, res) => {
   console.log('got req', req.headers, req.body);
