+++ server/tsconfig.json
@@ -2,7 +2,10 @@
   "compilerOptions": {
     "noImplicitAny": true,
     "strictNullChecks": true,
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "paths": {
+      "@shared/*": ["../src/shared/*"]
+    }
   },
   "include": ["."]
 }
